requirements.txt
openai
pinecone-client
python-dotenv
requests
feedparser
trafilatura
tqdm
unstructured
playwright
playwright-stealth
